chore: enable automatic trimming of whitespace in Markdown files
I tried finding why this was disabled to begin with; however, this
predates the project, and I am not sure the origin. My guess is
that this stems from the Markdown specification allowing two spaces
to force a hard line break. By trimming trailing whitespace, that
could mean that Markdown formatting is affected. However, we do not
use two spaces to force linebreaks. Instead, we use blank lines. As
such, I think we can enable automatic whitespace trimming.